fix: show professional locations
