+++ src/app/master/components/master-profile-info/master-profile-info.component.ts
@@ -1,11 +1,13 @@
 import {Component} from '@angular/core';
+import {ProfessionalLocationInline} from '@app/api/models/professional-location-inline';
 import {HelperService} from '@app/core/services/helper.service';
+import {FullLocationService} from '@app/core/services/location/full-location.service';
 import MasterProfileContext from '@app/master/interfaces/master-profile-context.interface';
 import {MasterProfileContextService} from '@app/master/services/master-profile-context.service';
 import {Language} from '@app/profile/models/language';
 import {LanguagesApiService} from '@app/profile/services/languages-api.service';
 import {UserLanguagesApiService} from '@app/profile/services/user-languages-api.service';
-import {Observable} from 'rxjs';
+import {forkJoin, Observable} from 'rxjs';
 import {first, map, shareReplay, switchMap} from 'rxjs/operators';
 
 function isNumbers(array: any[]): array is number[] {
@@ -21,11 +23,13 @@ export class MasterProfileInfoComponent {
 
     public context$: Observable<MasterProfileContext>;
     public languages$: Observable<Language[]>;
+    public locations$: Observable<{ id: number; text: string }[]>;
     public readonly editDefaultUrl = 'professional-contact-add-default/';
     public readonly editUrl = 'professional-contact-edit/';
     public readonly addUrl = 'professional-contact-add/';
 
     constructor(
+        private readonly fullLocationService: FullLocationService,
         context: MasterProfileContextService,
         userLanguagesApi: UserLanguagesApiService,
         languagesApi: LanguagesApiService
@@ -38,6 +42,11 @@ export class MasterProfileInfoComponent {
             switchMap(ids => userLanguagesApi.getList(ids)),
             switchMap(languages => languagesApi.getList(languages.map(lang => lang?.language))),
             shareReplay(1)
+        );
+        this.locations$ = this.context$.pipe(
+            switchMap(({master}) =>
+                forkJoin(master.locations.map(x => this.fullLocationService.getTextLocation(x)))
+            )
         );
     }
