+++ src/app/order/components/location-step/location-step.component.ts
@@ -1,13 +1,13 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, forwardRef } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
-import { ProfessionalLocationInline } from '@app/api/models';
-import { UserLocationApiService } from '@app/core/services';
-import { FullLocationService } from '@app/core/services/location/full-location.service';
-import { StepComponent } from '@app/order/abstract/step';
+import {ChangeDetectionStrategy, ChangeDetectorRef, Component, forwardRef} from '@angular/core';
+import {FormControl, Validators} from '@angular/forms';
+import {ProfessionalLocationInline} from '@app/api/models';
+import {UserLocationApiService} from '@app/core/services';
+import {FullLocationService} from '@app/core/services/location/full-location.service';
+import {StepComponent} from '@app/order/abstract/step';
 import LocationStepData from '@app/order/interfaces/location-step-data.interface';
 import StepContext from '@app/order/interfaces/step-context.interface';
-import { forkJoin, Observable } from 'rxjs';
-import { map, switchMap, takeUntil } from 'rxjs/operators';
+import {forkJoin} from 'rxjs';
+import {switchMap, takeUntil} from 'rxjs/operators';
 
 export enum LocationType {
     Online = 'online',
@@ -53,7 +53,7 @@ export class LocationStepComponent extends StepComponent<LocationStepData> {
 
     protected onContextChanged(context: StepContext): void {
         super.onContextChanged(context);
-        const { service = null } = context;
+        const {service = null} = context;
         switch (service.service_type) {
             case LocationType.Online: {
                 this.locationLabel = 'service-location.online';
@@ -83,8 +83,8 @@ export class LocationStepComponent extends StepComponent<LocationStepData> {
 
     // TODO Extract getting locations out of component
     private getServiceLocations(): void {
-        const locationsObservables = this.context.service.locations.map(({ location }) => {
-            return this.getTextLocation(location);
+        const locationsObservables = this.context.service.locations.map(({location}) => {
+            return this.fullLocationService.getTextLocation(location);
         });
         forkJoin(locationsObservables)
             .pipe(takeUntil(this.ngDestroy$))
@@ -96,30 +96,17 @@ export class LocationStepComponent extends StepComponent<LocationStepData> {
 
     private getClientLocations(): void {
         this.userLocationService.getByClientId().pipe(
-            switchMap(({ results: locations }) => {
-                return forkJoin(locations.map((location) => this.getTextLocation((location as unknown) as ProfessionalLocationInline)));
+            switchMap(({results: locations}) => {
+                return forkJoin(
+                    locations.map((location) =>
+                        this.fullLocationService.getTextLocation((location as unknown) as ProfessionalLocationInline))
+                );
             }),
             takeUntil(this.ngDestroy$)
         ).subscribe(locations => {
             this.locations = locations;
             this.cd.markForCheck();
         });
-    }
-
-    private getTextLocation(location: ProfessionalLocationInline): Observable<{ id: number; text: string }>  {
-        return this.fullLocationService.getFullLocation(location).pipe(
-            map(res => {
-                const textLocation = ['country', 'city']
-                    .map(key => res?.[key]?.name)
-                    .filter(value => Boolean(value))
-                    .join(', ');
-
-                return {
-                    id: location.id,
-                    text: `${textLocation}, ${location.address}`
-                };
-            })
-        );
     }
 
     private subscribeFormControl(): void {
@@ -132,9 +119,9 @@ export class LocationStepComponent extends StepComponent<LocationStepData> {
     private getStepState(): LocationStepData {
         return this.locationKey
             ? {
-                  ...initState,
-                  [this.locationKey]: this.formControl.value
-              }
+                ...initState,
+                [this.locationKey]: this.formControl.value
+            }
             : initState;
     }
 }
